% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap2r.R
\name{rdinar}
\alias{rdinar}
\title{rdinar}
\usage{
rdinar(xname)
}
\arguments{
\item{xname}{Variable name}
}
\value{
logical
}
\description{
Check whether variable is an array
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

