% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap2r.R
\name{rdindt}
\alias{rdindt}
\title{rdindt}
\usage{
rdindt(xname)
}
\arguments{
\item{xname}{Variable name}
}
\value{
character
}
\description{
Return data type of a variable
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

