% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap2r.R
\name{rdinit}
\alias{rdinit}
\title{rdinit}
\usage{
rdinit(datfil, iunit = 30, iulog = 0)
}
\arguments{
\item{datfil}{name of configuration file to read}

\item{iunit}{}

\item{iulog}{}
}
\value{
a \code{list} of formal arguments
}
\description{
Initializes TTUTIL for reading
}
\details{
Supply FORTRAN file descriptors
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

