\name{rdinlv}
\alias{rdinlv}
\title{rdinlv}
\usage{
  rdinlv(setflag = FALSE)
}
\arguments{
  \item{setflag}{Flag through which is indicated whether
  information from a possibly active rerun set is wanted}
}
\value{
  a vector of variables
}
\description{
  Return list of variables
}
\author{
  Mikhail Titov (\email{mlt@gmx.us})
}

