% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap2r.R
\name{rdinne}
\alias{rdinne}
\title{rdinne}
\usage{
rdinne(xname)
}
\arguments{
\item{xname}{Variable name}
}
\value{
# of elements
}
\description{
Return # of elements of a variable
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

