% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap2r.R
\name{rdread}
\alias{rdread}
\title{rdread}
\usage{
rdread(xname)
}
\arguments{
\item{xname}{parameter}
}
\value{
a value read
}
\description{
Universal read function
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

