% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap2r.R
\name{read.tt}
\alias{read.tt}
\title{read.tt}
\usage{
read.tt(fname)
}
\arguments{
\item{fname}{file name to read data from}
}
\value{
\code{list} with parameters read.
}
\description{
Read all parameters from the file into a list.
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

