% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{swap2r}
\alias{swap2r}
\alias{swap2r-package}
\title{Interface to SWAP hydrological model}
\description{
Utility functions to read SWAP output as well as read and
write its configuration files via TTutil Fortran library.
}
\details{
\tabular{ll}{
Version: \tab 0.1\cr
Licence: \tab GPLv3, also see LICENSE.txt for underlying TTutil\cr
}
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

