% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap2r.R
\name{wrclose}
\alias{wrclose}
\title{wrclose}
\usage{
wrclose(iunit = 50)
}
\arguments{
\item{iunit}{unit to close}
}
\value{
a \code{list} of formal arguments
}
\description{
Close fortrain unit
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

