\name{wrinit}
\alias{wrinit}
\title{wrinit}
\usage{
  wrinit(datfil, iunit = 50)
}
\arguments{
  \item{datfil}{name of configuration file to write}

  \item{iunit}{some unit number to associate}
}
\value{
  a \code{list} of formal arguments
}
\description{
  Initializes TTUTIL for writing
}
\details{
  Supply FORTRAN file descriptor
}
\author{
  Mikhail Titov (\email{mlt@gmx.us})
}

