\name{write.tt}
\alias{write.tt}
\title{write.tt}
\usage{
  write.tt(x, fname)
}
\arguments{
  \item{x}{list with parameters to write}

  \item{fname}{file name to read data from}
}
\value{
  Nothing
}
\description{
  Write list with parameters to a file.
}
\details{
  List will be flattened.
}
\author{
  Mikhail Titov (\email{mlt@gmx.us})
}

