% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap2r.R
\name{wrwrite}
\alias{wrwrite}
\title{wrwrite}
\usage{
wrwrite(xname, x)
}
\arguments{
\item{xname}{parameter}

\item{x}{value}
}
\value{
a \code{list} of formal arguments
}
\description{
Universal writes function
}
\author{
Mikhail Titov (\email{mlt@gmx.us})
}

